Promoted IRODSUriUtils.UserInfo to lop level
The new class is called IRODSUriUserInfo.  It contains enough logic that it should be at the top level.
