+++ jargon-core/src/test/java/org/irods/jargon/core/unittest/AllTests.java
@@ -6,7 +6,7 @@ import org.irods.jargon.core.query.TranslatedIRODSQueryTest;
 import org.irods.jargon.core.remoteexecute.RemoteExecuteServiceImplTest;
 import org.irods.jargon.core.security.IRODSPasswordUtilitiesTest;
 import org.irods.jargon.core.utils.IRODSUriUtilsTest;
-import org.irods.jargon.core.utils.IRODSUriUtilsUserInfoTest;
+import org.irods.jargon.core.utils.IRODSUriUserInfoTest;
 import org.irods.jargon.core.utils.LocalFileUtilsTest;
 import org.irods.jargon.core.utils.MiscIRODSUtilsTest;
 import org.junit.runner.RunWith;
@@ -19,7 +19,7 @@ import org.junit.runners.Suite;
 		PackingInstructionTests.class, DomainTests.class, TransferTests.class,
 		LocalFileUtilsTest.class, RemoteExecuteServiceImplTest.class,
 		IRODSPasswordUtilitiesTest.class, IRODSUriUtilsTest.class,
-		IRODSUriUtilsUserInfoTest.class, MiscIRODSUtilsTest.class,
+		IRODSUriUserInfoTest.class, MiscIRODSUtilsTest.class,
 		AuthTests.class, ChecksumTests.class, TransferRestartTests.class })
 /**
  * Suite to run all tests (except long running and functional), further refined by settings in testing.properites.  Some subtests may be shut
